+++ android/app/src/main/java/com/sellingcartapp/MainApplication.java
@@ -11,6 +11,8 @@ import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
+import com.sellingcartapp.bridge.SellingCartPackage;
+
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
